agent-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
requests==2.32.3
python-dotenv==1.0.1
typing-extensions==4.12.2
mysql-connector-python==8.2.0